implement shimloader support, add votv
Voices of the Void is a game about hunting aliens and the wacky /
spooky events that happen as a result.

Shimloader is my WIP modloader which shims the UE4SS loader and
a mod manager (R2MM, for example). It allows Unreal to load blueprint
mods (.pak files) and ue4ss mods (.lua) from external directories.

Blueprint and eu4ss mods are installed into the respective dirs:
- PROFILE/shimloader/bp_mods
- PROFILE/shimloader/ue4ss_mods
which are then injected into the game at runtime by shimloader.

The shimloader package can be found here:
/c/voices-of-the-void/p/0xFFF7/votv_shimloader/